Update ballistic_armor.json
Slightly adjusted the speed of drawing magazines from the pouches of filled plate carriers and slowed down the same from unfilled plate carriers. PC magazine pouches are fast to use IRL due to the proximity of the weapon and how little the hand has to move. Also, the plate's rigidity contributes to this ease of drawing, so unfilled PCs are less rigid and thus more likely to shift and slow down your draw.